Update index.html
Key Changes and Updates:
Firebase Firestore Configuration:

The Firebase configuration remains as you had it. If there are any issues with the storage bucket or Firestore setup, they’ll need to be fixed in the Firebase console.
Team Selection Form:

The form allows participants to input Player 1 and Player 2. Once submitted, the team is added to Firestore.
Displaying Teams:

The list of submitted teams will be fetched from Firestore, displaying the players' names as well as their current points.
Points Update:

A simple placeholder to update the points is added (via updateTeamPoints function). You would need to extend this to calculate points based on match performance, and this can be triggered after each match or update.